requirements.txt
aiogram>=3.0.0
python-dotenv
apscheduler==3.10.4
pypdf
